% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProps.R
\name{calcProps}
\alias{calcProps}
\title{A function to caluclate the proportion of cell types from a Seurat object}
\usage{
calcProps(data)
}
\arguments{
\item{data}{The Seurat object containing cell cluster identities}
}
\description{
This function takes as input a Seurat object, which contains cell types identified by the SingleR package. The function calculates the relative amount of each cell type.
}
\examples{
calcProps(data = my_seurat_object)
}
\keyword{Seurat}
\keyword{cell}
\keyword{proportions,}
\keyword{types,}
