% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePC.R
\name{choosePC}
\alias{choosePC}
\title{A function to choose principal components}
\usage{
choosePC(data, perms, plot, seed, center, scale)
}
\arguments{
\item{data}{The data on which PCA will be performed}

\item{perms}{How many permutations to perform on the data, default is 10}

\item{plot}{Whether or not a plot should be returned comparing the two PCA, default is FALSE}

\item{seed}{The seed to use for reshuffling dataframe columns, default is 629}

\item{center}{Whether or not data should be centered prior to PCA, default is TRUE}

\item{scale}{Whether or not data should be scaled prior to PCA, default is TRUE}
}
\description{
This function chooses an optimal number of principal components to explain the variance in your data. The amount of variance explained by the principal components is compared to the amount of variance explained by the principle components of a random reshuffing of the columns of the original dataframe, and the two are compared.
}
\examples{
choosePC(data = my_df, plot = TRUE, center = TRUE, scale = FALSE)
}
\keyword{PCA,}
\keyword{dimension}
\keyword{reduction}
\keyword{t-SNE,}
