% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefSummary.R
\name{CoefSummary}
\alias{CoefSummary}
\title{A function to provide a summary of a linear model.}
\usage{
CoefSummary(
  model = NULL,
  kable = TRUE,
  conf.level = 0.05,
  font = "Avenir",
  kable.caption = NULL,
  table.digits = 4
)
}
\arguments{
\item{model}{The model to be analyzed. Defaults to NULL.}

\item{kable}{Should the table be returned as an interactive kable instead of a simple dataframe? Defaults to TRUE.}

\item{conf.level}{The confidence level for determining variable significance. Defaults to 95% aka 0.05.}

\item{font}{What font should be used? Defaults to "Avenir".}

\item{kable.caption}{Define a caption to be placed above the table. Defaults to "Model Coefficients".}

\item{table.digits}{The number of digits to round to. Defaults to 4.}
}
\description{
This function takes in the output from lm() or glm() and provides a tidy, aesthetically-pleasing summary of the coefficients.
}
\examples{
CoefSummary(model = my_glm, font = "Georgia")
}
