% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CooksD.R
\name{CooksD}
\alias{CooksD}
\title{Calculate Cook's Distance.}
\usage{
CooksD(leverage, level, model, positive)
}
\arguments{
\item{leverage}{The leverage of each point.}

\item{level}{The level above which an observation is considered "high-leverage."}

\item{model}{The model which you'd like to analyze.}

\item{positive}{Should positive or negative Cook's Distance values be computed?}
}
\description{
A helper function to calculate Cook's Distance outliers in GLMs. Meant to be called from CooksDPlot().
}
\examples{
CooksD(leverage = ., level = .5, model = my_lm, positive = TRUE)
}
