% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractLegend.R
\name{ExtractLegend}
\alias{ExtractLegend}
\title{Extract the legend from a ggplot2 object.}
\usage{
ExtractLegend(my.plot = NULL)
}
\arguments{
\item{my.plot}{The plot from which to extract the legend. Defaults to NULL.}
}
\description{
This function pulls out the legend box from a ggplot2 object, which you can then arrange with other objects using e.g. patchwork.
}
\examples{
ExtractLegend(plot1)
}
