% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAnalyze.R
\name{ggAnalyze}
\alias{ggAnalyze}
\title{A function to analyze a linear model visually}
\usage{
ggAnalyze(mod)
}
\arguments{
\item{mod}{This is the model to be analyzed}
}
\description{
This function returns the four plots returned by default plot.lm(), but using ggplot2
}
\examples{
ggAnalyze(mod = my_lm)
}
\keyword{ggplot2}
\keyword{linear,}
\keyword{model,}
