% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negateIn.R
\name{\%!in\%}
\alias{\%!in\%}
\title{A function to negate the %in% operator.}
\usage{
... \%!in\% NA
}
\description{
This function returns the complement / inverse of the %in% operator. Useful for subsetting, etc.
}
\examples{
df[df$var \%!in\% pattern, ]
}
