% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{Implementing the jackknife method of estimating standard errors}
\usage{
jackknife(vx, tfunc, dfunc, im, vw, ...)
}
\arguments{
\item{vx}{a vector of observations}

\item{tfunc}{the function which takes the rho argument and returns the statistic}

\item{dfunc}{the delta function which takes a tuple or window of the observations and returns a vector of deltas}

\item{im}{the tuple or window size}

\item{vw}{the weights for deleting or downweighting the blocks}

\item{...}{other arguments for both \code{tfunc} and \code{dfunc} if any}
}
\value{
a list containing the results

The object is a list containing the following components:
\item{T}{the value of the statistic}
\item{jack}{a vector or matrix containing the jackknifed statistics}
\item{sig2}{the jackknifed estimate of the variance of the statistic}
}
\description{
Implementing the jackknife method of estimating standard errors of some linear statistics
}
\details{
The function takes several arguments including functions and returns a list containing the results.
}
\examples{
vw1 = 1
vw2 = c(0.25,0.75,1,0.75,0.25)

# the AR(1) model
iN = 100
vx <- rnorm(iN)

# for first order autocorrelation
# dfunc vec 2, y_t and y_t-1
dfunc <- function(yy, my, vy) return(prod(yy - my)/vy)

tfunc <- function(rho, my, vy) return(rho)

# results
ret = jackknife(as.vector(vx), tfunc, dfunc, im=2, vw=vw1, my=mean(vx), vy=var(vx))
ret

ret = jackknife(as.vector(vx), tfunc, dfunc, im=2, vw=vw2, my=mean(vx), vy=var(vx))
ret

}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
\keyword{jackknife}
