% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_jack.R
\name{legend_jack}
\alias{legend_jack}
\title{A function to control the legend.}
\usage{
legend_jack(legend.rows = 1, key.size = 3)
}
\arguments{
\item{legend.rows}{How many rows should the legend have? Mostly useful for top/bottom legends. Defaults to 1.}

\item{key.size}{How large should the legend icons be? Defaults to 3.}
}
\description{
This function will serve as an easier-to-use wrapper around guide_legend().
}
\examples{
legend_jack(nrow = 2)
}
