% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelDiagnostics.R
\name{ModelDiagnostics}
\alias{ModelDiagnostics}
\title{A function to provide model diagnostic statistics.}
\usage{
ModelDiagnostics(
  model = NULL,
  kable = TRUE,
  kable.font = "Avenir",
  kable.caption = NULL,
  kable.digits = 4
)
}
\arguments{
\item{model}{The model to be analyzed. Defaults to NULL.}

\item{kable}{Should the output be placed in kable format, or returned as a simple dataframe? Defaults to TRUE.}

\item{kable.font}{What font should be used? Defaults to "Avenir".}

\item{kable.caption}{Define a caption to be placed above the table. Defaults to "Model Coefficients".}

\item{kable.digits}{The number of digits to round to. Defaults to 4.}
}
\description{
This function builds on the output of glance() in order to make it neater and more aesthetically pleasing.
}
\examples{
ModelDiagnostics(my_glm, font = "Comic Sans")
}
