% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPlot.R
\name{OutlierPlot}
\alias{OutlierPlot}
\title{Generate an outlier plot.}
\usage{
OutlierPlot(model = NULL)
}
\arguments{
\item{model}{The model you'd like to analyze. Defaults to NULL.}
}
\description{
Reimplements the standardized residuals vs. leverage plot from base R, along with the Cook's Distance cutoff lines for outliers, in ggplot2. Points with leverage values greater than 2p/n are colored blue.
}
\examples{
OutlierPlot(model = my_lm)
}
