% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQPlot.R
\name{QQPlot}
\alias{QQPlot}
\title{Generate a QQ-plot.}
\usage{
QQPlot(model = NULL)
}
\arguments{
\item{model}{The model you'd like to analyze. Defaults to NULL.}
}
\description{
This function reimplements the base R QQ-plot using ggplot2.
}
\examples{
QQPlot(model = my_lm)
}
