% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderRmd.R
\name{renderRMD}
\alias{renderRMD}
\title{An RMarkdown function}
\usage{
renderRMD()
}
\description{
This function renders an RMarkdown file that performs computatinoally intensive single cell RNAseq analysis on the longleaf HPC cluster.
}
\examples{
renderRMD()
}
\keyword{longleaf}
