% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlots.R
\name{savePlots}
\alias{savePlots}
\title{A Plot Saving Function}
\usage{
savePlots(plotList, geneList, wd)
}
\arguments{
\item{plotList}{A list of previously generated plots}

\item{geneList}{A list of gene names (or other labels) to use to name files}

\item{wd}{The directory to which you wish to save the images}
}
\description{
This function saves a list of plots of gene expression as .png files in a user-provided directory
}
\examples{
savePlots(plotList = myPlots, geneList = myGenes, wd = "~/username/plots")
}
