% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleLocationPlot.R
\name{ScaleLocationPlot}
\alias{ScaleLocationPlot}
\title{Generate a Scale-Location plot.}
\usage{
ScaleLocationPlot(model = NULL)
}
\description{
Reimplements the fitted values vs. square root-transformed standardized residuals plot found in plot.lm() in ggplot2.
}
\examples{
ScaleLocationPlot(model = my_lm)
}
