% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jack.R
\name{theme_jack}
\alias{theme_jack}
\title{A function for easily modifying themes.}
\usage{
theme_jack(
  grid = FALSE,
  legend = TRUE,
  border = TRUE,
  font = NULL,
  ticks = FALSE,
  axis.nums = TRUE,
  legend.pos = "bottom",
  legend.rows = 1,
  center.title = TRUE,
  y.axis.horiz = TRUE
)
}
\arguments{
\item{grid}{Should the plot include gridlines? Defaults to FALSE.}

\item{legend}{Should the legend be drawn? Defaults to TRUE.}

\item{border}{Should a rectangular border be drawn around the plot? Defaults to TRUE.}

\item{font}{A string detailing which font should be used on all plot text. Must be capitalized correctly. Defaults to NULL.}

\item{axis.nums}{Should the axes show number labels? Defaults to TRUE.}

\item{legend.pos}{Where should the legend be located? One of "left", "right", "bottom", "top". Defaults to "bottom".}

\item{legend.rows}{How many rows shows the legend have? Mostly relevant for top/bottom-located legends. Defaults to 1.}

\item{center.title}{Should the plot title and subtitle be centered? Defaults to TRUE.}

\item{y.axis.horiz}{Should the y-axis be horizontal in order to make reading easier? Defaults to TRUE.}
}
\description{
This function allows easy modification of a minimalist `ggplot2` theme - remove the grid, change fonts, etc.
}
\examples{
theme_jack(grid = TRUE, font = "Helvetica")
}
