% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeYehLab.R
\name{themeYehLab}
\alias{themeYehLab}
\title{A function for cleaner plots.}
\usage{
themeYehLab()
}
\description{
This function is a `ggplot2` theme designed with dimension reduction plots (think PCA, t-SNE, UMAP) in mind, but can be applied to other types of plots as needed. It removes axis ticks and text, and changes the font to Helvetica.
}
\examples{
themeYehLab()
}
