% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAnalyze.R
\name{ggAnalyze}
\alias{ggAnalyze}
\title{A function to analyze a linear model visually}
\usage{
ggAnalyze(mod)
}
\arguments{
\item{mod}{This is the model to be analyzed}
}
\description{
This function is a reimplementation of the four plots returned by default plot.lm() using the ggplot2 library.
}
\examples{
ggAnalyze(mod = my_lm)
}
