% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichmentScores.R
\name{plotES}
\alias{plotES}
\title{A function for plotting enrichment scores after GSEA}
\usage{
plotES(geneSetList, numPlots, fgseaRes, ranks)
}
\arguments{
\item{numPlots}{The number of plots you wish to generate}

\item{fgseaRes}{The results obtained from runnning fgsea()}

\item{ranks}{The ranks obtained from running fgsea()}

\item{geneSetLists}{A list of gene sets}
}
\description{
This function returns a list of enrichment plots after running GSEA using the 'fgsea' package
}
\examples{
plotES(geneSetList = myGeneList, numPlots = 5, fgseaRes = myResults, ranks = myRanks)
}
\keyword{enrichment}
