% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlots.R
\name{savePlots}
\alias{savePlots}
\title{A Plot Saving Function}
\usage{
savePlots(plot.list, gene.list, wd)
}
\arguments{
\item{wd}{The directory to which you wish to save the images}

\item{plotList}{A list of previously generated plots}

\item{geneList}{A list of gene names (or other labels) to use as filenames}
}
\description{
This function saves a list of plots of gene expression as .png files in a user-provided directory
}
\examples{
savePlots(plot.list = my_plots, gene.list = myy_enes, wd = "~/username/plots")
}
